Errores + Refactorizacion

+++ microservices/api/src/services/userService.js
@@ -1,43 +1,52 @@
 const DAOFactory = require("../dao/DAOFactory")
 const userDAO = DAOFactory.getUserDAO()
+const { NotFoundError } = require("../errors")
 
-async function findOrCreateUser({ keycloakId, email, name }) {
+class UserService {
 
-  let user = await userDAO.getUserByKeycloakId(keycloakId)
-  if (user) return user
+  async findOrCreateUser({ keycloakId, email, name }) {
+    let user = await userDAO.getUserByKeycloakId(keycloakId)
+    if (user) return user
 
-  user = await userDAO.getUserByEmail(email)
-  if (user) {
-    return await userDAO.updateUser(user.id, { name, email })
+    user = await userDAO.getUserByEmail(email)
+    if (user) {
+      return await userDAO.updateUser(user.id, { name, email })
+    }
+
+    return await userDAO.createUser({
+      keycloakId,
+      email,
+      name
+    })
   }
 
-  return await userDAO.createUser({
-    keycloakId,
-    email,
-    name
-  })
+  async getAllUsers() {
+    return userDAO.getAllUsers()
+  }
+
+  async getUserById(id) {
+    const user = await userDAO.getUserById(id)
+    if (!user) {
+      throw new NotFoundError("User not found")
+    }
+    return user
+  }
+
+  async updateUser(keycloakId, data) {
+    const user = await userDAO.getUserByKeycloakId(keycloakId)
+    if (!user) {
+      throw new NotFoundError("User not found")
+    }
+    return userDAO.updateUser(user.id, data)
+  }
+
+  async deleteUser(id) {
+    const user = await userDAO.getUserById(id)
+    if (!user) {
+      throw new NotFoundError("User not found")
+    }
+    return userDAO.deleteUser(id)
+  }
 }
 
-async function getAllUsers() {
-  return userDAO.getAllUsers()
-}
-
-async function getUserById(id) {
-  return userDAO.getUserById(id)
-}
-
-async function updateUser(keycloakId, data) {
-  return userDAO.updateUser(keycloakId, data)
-}
-
-async function deleteUser(id) {
-  return userDAO.deleteUser(id)
-}
-
-module.exports = {
-  findOrCreateUser,
-  getAllUsers,
-  getUserById,
-  updateUser,
-  deleteUser
-}
+module.exports = new UserService()
